task.md 23.3 steps 1 to 3: the butterfly history becomes signed, the gravity update replaces the overflow rescale, and the malus is switched on

+++ src/move_gen_service.rs
@@ -462,8 +462,12 @@ impl MoveGenService {
                     // `white_turn` is this node's side to move, i.e. the side that would play
                     // the move being ranked, which is the side the cutoff credited one ply up.
                     let side = crate::model::history_side(white_turn);
+                    // `task.md` 23.3: the entry is signed and bounded by `MAX_HISTORY`, so a
+                    // refuted quiet ranks *below* `BAND_QUIET` and an unseen one sits on it. The
+                    // band is 2,000,000 wide below the killers and `MAX_HISTORY` is 16,384, so
+                    // nothing can reach the band beneath.
                     crate::model::BAND_QUIET
-                        + unsafe { (*context.history_table)[side][from][to] } as i32
+                        + unsafe { (*context.history_table)[side][from][to] }
                 }
             };
 
@@ -1367,7 +1371,7 @@ mod tests {
         let zobrist_table = ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -1393,7 +1397,7 @@ mod tests {
         let zobrist_table = ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -1666,7 +1670,7 @@ mod tests {
         let zobrist_table = ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = Mutex::new(HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -1810,7 +1814,7 @@ mod tests {
 
         let zobrist_table = ZobristTable::with_capacity(1_000);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -1842,7 +1846,7 @@ mod tests {
         let zobrist_table = ZobristTable::with_capacity(1_000);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -1999,7 +2003,7 @@ mod tests {
         let zobrist_table = ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -2088,7 +2092,7 @@ mod tests {
                 let zobrist_table = ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
@@ -2138,7 +2142,7 @@ mod tests {
         let zobrist_table = crate::zobrist::ZobristTable::with_capacity(1);
         let stop_flag = std::sync::atomic::AtomicBool::new(false);
         let pv_nodes = std::sync::Mutex::new(std::collections::HashMap::new());
-        let history_table = [[[0u32; 64]; 64]; 2];
+        let history_table = [[[0i32; 64]; 64]; 2];
         let context = crate::model::SearchContext {
             zobrist_table: &zobrist_table,
             stop_flag: &stop_flag,
